Fix inlining of 'c' in efdt{Int,Word}{Up,Dn}FB
Summary:
By only having a single occurrence of 'c' in fusion helpers, we
guarantee the argument gets inlined once the helper has been
inlined. It appears that 'eft{Int,Word}' variants already were
implemented with this in mind, so this just brings 'enumFromThenTo'
on par.

This fixes #8763, where a loop body was too huge to get inlined.

Reviewers: hvr, bgamari, simonpj, nomeata

Subscribers: thomie, carter

GHC Trac Issues: #8763

Differential Revision: https://phabricator.haskell.org/D4547