Move as assertion, avoiding a crash in Data.Type.Equality
Prior to the LLF refactoring, the assertion was only attached
to the LevelledExpr component, probably so that the free
variable caching made possible through the first component was
still intact. This change shifts the check for coercions to
`lookupVar`, the only place where we construct the LevelledExpr
out of the list of variables in the type application. It
should be sufficient to check for Ids just there.