Respect RULEs and Unfoldings in ids to abstract over
There were rare occasions where the RHS of a
binding's RULE introduced by SpecConstr jumped to a
join point which would otherwise be dead.
Previously, those weren't present in the FIIs but
in the set of free variables, so that the floated
binding would spoil that outer join binding,
resulting in a panic of the Simplifier.

This commit brings the logic more in line with what
CoreFVs does.