Stabilise unfoldings only when we lifted something
Previously, pre-tidy stabilisation lead to huge (+62%)
regressions in veritas' Edlib.lhs, where the unfoldings
of $w/:>/ and $w/>/ were probably too huge to get inlined.

This patch moves the relevant stabilisation logic to FloatOut
where we can see if we actually floated out a binding. If
not, we keep the vanilla inlining.