Add RubbishLit for absent bindings of UnliftedRep
Trac #9279 reminded us that the worker wrapper transformation copes
really badly with absent unlifted bindings other than of type `Addr#`.

As `Note [Absent errors]` in WwLib.hs points out, we can't just use
`absentError` for unlifted bindings because there is no bottom to hide
the error in.
So instead, we synthesise a new `RubbishLit` of type
`forall (a :: TYPE 'UnliftedRep). a`, which code-gen may subsitute for
any pointer value. We choose `NULL`, so that there is a good chance that
the program crashes instead instead of leading to corrupt data, should
absence analysis have been too optimistic (#11126).